pyproject.toml
[project]
name = "mcp-video-transcriber"
version = "0.1.0"
description = "MCP server for transcribing videos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.12",
    "cryptography>=45.0.4",
    "fastapi>=0.115.0",
    "fastmcp>=2.8.0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
